txn: move draining_txs from tm_stm to tx_hash_ranges
Refactoring, nothing changes
