+++ src/v/cluster/tx_hash_ranges.h
@@ -153,6 +153,31 @@ struct tx_hash_ranges_set
 
 using repartitioning_id = named_type<int64_t, struct repartitioning_id_type>;
 
+struct draining_txs
+  : serde::envelope<draining_txs, serde::version<0>, serde::compat_version<0>> {
+    repartitioning_id id;
+    tx_hash_ranges_set ranges{};
+    absl::btree_set<kafka::transactional_id> transactions{};
+
+    draining_txs() = default;
+
+    draining_txs(
+      repartitioning_id id,
+      tx_hash_ranges_set ranges,
+      absl::btree_set<kafka::transactional_id> txs)
+      : id(id)
+      , ranges(std::move(ranges))
+      , transactions(std::move(txs)) {}
+
+    auto serde_fields() { return std::tie(id, ranges, transactions); }
+
+    bool contains(kafka::transactional_id tx_id) {
+        return transactions.contains(tx_id);
+    }
+
+    bool contains(tx_hash_type hash) { return ranges.contains(hash); }
+};
+
 namespace hosted_transactions {
 template<class T>
 tx_hash_ranges_errc
@@ -255,6 +280,21 @@ struct adl<cluster::tx_hash_ranges_set> {
         auto ranges
           = reflection::adl<std::vector<cluster::tx_hash_range>>{}.from(in);
         return {std::move(ranges)};
+    }
+};
+
+template<>
+struct adl<cluster::draining_txs> {
+    void to(iobuf& out, cluster::draining_txs&& dr) {
+        reflection::serialize(out, dr.id, dr.ranges, dr.transactions);
+    }
+    cluster::draining_txs from(iobuf_parser& in) {
+        auto id = reflection::adl<cluster::repartitioning_id>{}.from(in);
+        auto ranges
+          = reflection::adl<std::vector<cluster::tx_hash_range>>{}.from(in);
+        auto txs = reflection::adl<absl::btree_set<kafka::transactional_id>>{}
+                     .from(in);
+        return {id, std::move(ranges), std::move(txs)};
     }
 };
